And actually use the new BagReplicatorWorker
