+++ bag_replicator/src/main/scala/uk/ac/wellcome/platform/archive/bagreplicator/Main.scala
@@ -3,20 +3,14 @@ package uk.ac.wellcome.platform.archive.bagreplicator
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import com.amazonaws.services.s3.AmazonS3
+import com.amazonaws.services.sqs.AmazonSQSAsync
 import com.typesafe.config.Config
-import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.messaging.typesafe.NotificationStreamBuilder
+import uk.ac.wellcome.messaging.typesafe.{CloudwatchMonitoringClientBuilder, SQSBuilder}
+import uk.ac.wellcome.messaging.worker.monitoring.CloudwatchMonitoringClient
 import uk.ac.wellcome.platform.archive.bagreplicator.config.ReplicatorDestinationConfig
-import uk.ac.wellcome.platform.archive.bagreplicator.services.{
-  BagReplicator,
-  BagReplicatorWorker
-}
-import uk.ac.wellcome.platform.archive.common.config.builders.{
-  DiagnosticReporterBuilder,
-  IngestUpdaterBuilder,
-  OutgoingPublisherBuilder
-}
-import uk.ac.wellcome.platform.archive.common.ingests.models.BagRequest
+import uk.ac.wellcome.platform.archive.bagreplicator.services.{BagReplicator, BagReplicatorWorker}
+import uk.ac.wellcome.platform.archive.bagunpacker.config.builders.AlpakkaSqsWorkerConfigBuilder
+import uk.ac.wellcome.platform.archive.common.config.builders.{IngestUpdaterBuilder, OutgoingPublisherBuilder}
 import uk.ac.wellcome.storage.typesafe.S3Builder
 import uk.ac.wellcome.typesafe.WellcomeTypesafeApp
 import uk.ac.wellcome.typesafe.config.builders.AkkaBuilder
@@ -33,18 +27,22 @@ object Main extends WellcomeTypesafeApp {
 
     implicit val s3Client: AmazonS3 = S3Builder.buildS3Client(config)
 
+    implicit val monitoringClient: CloudwatchMonitoringClient =
+      CloudwatchMonitoringClientBuilder.buildCloudwatchMonitoringClient(config)
+
+    implicit val sqsClient: AmazonSQSAsync =
+      SQSBuilder.buildSQSAsyncClient(config)
+
     val operationName = "replicating"
 
     new BagReplicatorWorker(
-      stream = NotificationStreamBuilder
-        .buildStream[BagRequest](config),
-      ingestUpdater = IngestUpdaterBuilder.build(config, operationName),
-      outgoing = OutgoingPublisherBuilder.build(config, operationName),
-      reporter = DiagnosticReporterBuilder.build(config),
-      replicator = new BagReplicator(
+      alpakkaSQSWorkerConfig = AlpakkaSqsWorkerConfigBuilder.build(config),
+      bagReplicator = new BagReplicator(
         config = ReplicatorDestinationConfig
           .buildDestinationConfig(config)
-      )
+      ),
+      ingestUpdater = IngestUpdaterBuilder.build(config, operationName),
+      outgoingPublisher = OutgoingPublisherBuilder.build(config, operationName)
     )
   }
 }
